Language.PureScript.Interpreter

High-level API

buildCoreFn :: MonadFix m => Env m -> Module ann -> EvalT m (Env m)

Evaluate each of the bindings in a compiled PureScript module, and store the evaluated values in the environment, without evaluating any main expression.

builtIn :: ToValue m a => ModuleName -> Text -> a -> Env m

Create an environment from a Haskell value.

It is recommended that a type annotation is given for the type of the value being provided.

For example:

   builtIn "greeting" ("Hello, World!" :: Text)
   builtIn "somePrimes" ([2, 3, 5, 7, 11] :: Vector Integer)

Functions can be provided as built-ins, but the EvalT monad needs to be used to wrap any outputs (or values in positive position):

   builtIn "strip" ((pure . Text.strip) :: Text -> EvalT m Text)
   builtIn "map" (traverse :: (Value -> EvalT m Value) -> Vector Value -> EvalT m (Vector Value))

Polymorphic functions can also be provided as built-ins, but values with polymoprhic types will need to be passed across the FFI boundary with monomorphic types. The type Value can always be used to represent values of unknown or polymorphic type, as in the map example above.

Evaluation

Representation of values

data Value m

The representation of values used by the interpreter - essentially, the semantic domain for a simple untyped lambda calculus with records and ADTs.

Any additional side effects which might occur in FFI calls to Haskell code are tracked by a monad in the type argument.

Constructors

Object (HashMap Text (Value m))
Records are represented as hashmaps from their field names to values

Array (Vector (Value m))
Arrays

String Text

Number Scientific
Numeric values

TODO: separate integers from floating-point values

Bool Bool

Closure (Value m -> EvalT m (Value m))
Closures, represented in higher-order abstract syntax style.

Constructor (Qualified (ProperName 'ConstructorName)) [Value m]
Fully-applied data constructors

Evaluation monad

newtype EvalT m a

The monad used by the interpreter, which supports error reporting for errors which can occur during evaluation.

The transformed monad is used to track any benign side effects that might be exposed via the foreign function interface to PureScript code.

Constructors

EvalT

Fields

unEvalT :: ExceptT EvaluationError m a

runEvalT :: EvalT m a -> m (Either EvaluationError a)

type Eval = EvalT Identity

Non-transformer version of EvalT, useful in any settings where the FFI does not use any side effects during evaluation.

runEval :: Eval a -> Either EvaluationError a

data EvaluationError

Errors which can occur during evaluation of PureScript code.

PureScript is a typed language, and tries to prevent runtime errors. However, in the context of this interpreter, we can receive data from outside PureScript code, so it is possible that runtime errors can occur if we are not careful. This is similar to how PureScript code can fail at runtime due to errors in the FFI.

Constructors

UnknownIdent (Qualified Ident)
A name was not found in the environment

TypeMismatch Text
The runtime representation of a value did not match the expected representation

FieldNotFound Text
A record field did not exist in an Object value.

InexhaustivePatternMatch
A pattern match failed to match its argument

InvalidNumberOfArguments Int Int
A pattern match received the wrong number of arguments

UnsaturatedConstructorApplication
A pattern match occurred against a partially-applied data constructor

InvalidFieldName PSString
A PureScript string which contains lone surrogates which could not be decoded. See PSString.

OtherError Text
An error occurred in a foreign function which is not tracked by any of the other error types.

TODO: remove this in favor of using monadic FFI functions

renderEvaluationError :: EvaluationError -> String

Render an EvaluationError as a human-readable string.

Eval/apply

type Env m = Map (Qualified Ident) (Value m)

An environment, i.e. a mapping from names to evaluated values.

An environment for a single built-in function can be constructed using the builtIn function, and environments can be combined easily using the Monoid instance for Map.

eval :: forall m. MonadFix m => ModuleName -> Env m -> Expr () -> EvalT m (Value m)

Evaluate a PureScript CoreFn expression in the given environment.

Note: it should not be necessary to call this function directly in most circumstances. It is provided as a helper function, for some more advanced use cases, such as setting up a custom environment.

apply :: MonadFix m => Value m -> Value m -> EvalT m (Value m)

Apply a value which represents an unevaluated closure to an argument.

Foreign function interface

data FFI m

Constructors

FFI

Fields

ffi_moduleName :: ModuleName

ffi_values :: [ForeignImport m]

data ForeignImport m

Constructors

ForeignImport

Fields

fv_name :: Ident

fv_type :: SourceType

fv_value :: Value m

toEnv :: FFI m -> Env m

toExterns :: FFI m -> ExternsFile

Conversion to and from Haskell types

class MonadFix m => ToValue m a where

Values which can be communicated across the FFI boundary from Haskell to PureScript.

Instances should identify and document any valid representations as a subset of the semantic domain Value. Such a subset can be identified by an injective function toValue, and a partial inverse, fromValue, defined on the image of toValue.

Laws:

   fromValue . toValue = pure

Minimal complete definition

Nothing

Methods

toValue :: a -> Value m

The default implementation uses generic deriving to identify a Haskell record type with a single data constructor with a PureScript record with the same field names.

default toValue :: (Generic a, ToObject m (Rep a)) => a -> Value m

fromValue :: Value m -> EvalT m a

default fromValue :: (Generic a, ToObject m (Rep a)) => Value m -> EvalT m a

Higher-order functions

class ToValueRHS m a where

ToValue should support functions with types such as

   a -> EvalT m b
   a -> b -> EvalT m c
   a -> b -> c -> EvalT m d
   (a -> EvalT m b) -> EvalT m c
   (a -> b -> EvalT m c) -> EvalT m d

Note that every type in a return position is wrapped in the EvalT monad transformer. This is because evaluation in general may result in errors. However, a naive translation would result in too many applications of EvalT.

Specifically, we do not want to require types such as these, in which EvalT appears on the right hand side of every function arrow:

   a -> EvalT m b (b -> EvalT m c)
   a -> EvalT m b (b -> EvalT m (c -> EvalT m d))

For this reason, the ToValue instance for functions delegates to this type class for the type on the right hand side of the function. It skips the application of EvalT for nested function types.

Methods

toValueRHS :: a -> EvalT m (Value m)

fromValueRHS :: EvalT m (Value m) -> a

Records

data ObjectOptions

Options for customizing generic deriving of record instances

Constructors

ObjectOptions

Fields

toPureScriptField :: Text -> Text

Map a Haskell field name to a PureScript field name on the corresponding record type.

defaultObjectOptions :: ObjectOptions

Maps Haskell field names to PureScript field names, unmodified.

genericToValue :: (MonadFix m, Generic a, ToObject m (Rep a)) => ObjectOptions -> a -> Value m

Derived toValue function for Haskell record types which should map to corresponding PureScript record types.

genericFromValue :: (MonadFix m, Generic a, ToObject m (Rep a)) => ObjectOptions -> Value m -> EvalT m a

Derived fromValue function for Haskell record types which should map to corresponding PureScript record types.

class ToObject m f where

This class is used in the default instance for ToValue, via generic deriving, in order to identify a Haskell record type (with a single data constructor and named fields) with values in the semantic domain corresponding to a PureScript record type with the same field names.

Methods

toObject :: ObjectOptions -> f x -> HashMap Text (Value m)

fromObject :: ObjectOptions -> HashMap Text (Value m) -> EvalT m (f x)