Language.PureScript.Interpreter.Monad

High-level API

data InterpretT m a

runInterpretT :: Monad m => InterpretT m a -> m (Either InterpretError a)

runInterpret :: Interpret a -> Either InterpretError a

Error messages

data InterpretError

Constructors

EvaluationError EvaluationError

BuildError BuildError

renderInterpretError :: InterpretError -> String

Foreign function interface

ffi :: Monad m => FFI m -> InterpretT m ()

Building PureScript source

build :: MonadFix m => Text -> InterpretT m (Module Ann)

buildCoreFn :: MonadFix m => Module Ann -> InterpretT m (Module Ann)

Evaluating values

eval :: (MonadFix m, ToValueRHS m a) => Expr () -> InterpretT m a

evalMain :: (MonadFix m, ToValueRHS m a) => ModuleName -> InterpretT m a

Re-exports

class Monad m => MonadFix (m :: Type -> Type)

Monads having fixed points with a 'knot-tying' semantics. Instances of MonadFix should satisfy the following laws:

Purity

mfix (return . h) = return (fix h)

Left shrinking (or Tightening)

mfix (\x -> a >>= \y -> f x y) = a >>= \y -> mfix (\x -> f x y)

Sliding

mfix (liftM h . f) = liftM h (mfix (f . h)), for strict h.

Nesting

mfix (\x -> mfix (\y -> f x y)) = mfix (\x -> f x x)

This class is used in the translation of the recursive do notation supported by GHC and Hugs.

Minimal complete definition

mfix

newtype ModuleName

Constructors

ModuleName Text

type Ann = (SourceSpan, [Comment], Maybe SourceType, Maybe Meta)

data Expr a

data Module a

data EvaluationError

Errors which can occur during evaluation of PureScript code.

PureScript is a typed language, and tries to prevent runtime errors. However, in the context of this interpreter, we can receive data from outside PureScript code, so it is possible that runtime errors can occur if we are not careful. This is similar to how PureScript code can fail at runtime due to errors in the FFI.

Constructors

UnknownIdent (Qualified Ident)
A name was not found in the environment

TypeMismatch Text
The runtime representation of a value did not match the expected representation

FieldNotFound Text
A record field did not exist in an Object value.

InexhaustivePatternMatch
A pattern match failed to match its argument

InvalidNumberOfArguments Int Int
A pattern match received the wrong number of arguments

UnsaturatedConstructorApplication
A pattern match occurred against a partially-applied data constructor

InvalidFieldName PSString
A PureScript string which contains lone surrogates which could not be decoded. See PSString.

OtherError Text
An error occurred in a foreign function which is not tracked by any of the other error types.

TODO: remove this in favor of using monadic FFI functions

data BuildError

Constructors

UnableToParse (NonEmpty ParserError)

UnableToCompile MultipleErrors