Language.PureScript.Interpreter.Types

Evaluation

Value types

data Value m

The representation of values used by the interpreter - essentially, the semantic domain for a simple untyped lambda calculus with records and ADTs.

Any additional side effects which might occur in FFI calls to Haskell code are tracked by a monad in the type argument.

Constructors

Object (HashMap Text (Value m))
Records are represented as hashmaps from their field names to values

Array (Vector (Value m))
Arrays

String Text

Number Scientific
Numeric values

TODO: separate integers from floating-point values

Bool Bool

Closure (Value m -> EvalT m (Value m))
Closures, represented in higher-order abstract syntax style.

Constructor (ProperName 'ConstructorName) [Value m]
Fully-applied data constructors

Evaluation monad

type Env m = Map (Qualified Ident) (Value m)

An environment, i.e. a mapping from names to evaluated values.

An environment for a single built-in function can be constructed using the builtIn function, and environments can be combined easily using the Monoid instance for Map.

newtype EvalT m a

The monad used by the interpreter, which supports error reporting for errors which can occur during evaluation.

The transformed monad is used to track any benign side effects that might be exposed via the foreign function interface to PureScript code.

Constructors

EvalT

Fields

unEvalT :: ExceptT EvaluationError m a

runEvalT :: EvalT m a -> m (Either EvaluationError a)

type Eval = EvalT Identity

Non-transformer version of EvalT, useful in any settings where the FFI does not use any side effects during evaluation.

runEval :: Eval a -> Either EvaluationError a

Evaluation errors

data EvaluationError

Errors which can occur during evaluation of PureScript code.

PureScript is a typed language, and tries to prevent runtime errors. However, in the context of this interpreter, we can receive data from outside PureScript code, so it is possible that runtime errors can occur if we are not careful. This is similar to how PureScript code can fail at runtime due to errors in the FFI.

Constructors

UnknownIdent (Qualified Ident)
A name was not found in the environment

TypeMismatch Text
The runtime representation of a value did not match the expected representation

FieldNotFound Text
A record field did not exist in an Object value.

InexhaustivePatternMatch
A pattern match failed to match its argument

InvalidNumberOfArguments Int Int
A pattern match received the wrong number of arguments

UnsaturatedConstructorApplication
A pattern match occurred against a partially-applied data constructor

InvalidFieldName PSString
A PureScript string which contains lone surrogates which could not be decoded. See PSString.

OtherError Text
An error occurred in a foreign function which is not tracked by any of the other error types.

TODO: remove this in favor of using monadic FFI functions

renderEvaluationError :: EvaluationError -> String

Render an EvaluationError as a human-readable string.