Language.PureScript.Make.Simplified

Documentation

data BuildError

Constructors

UnableToParse (NonEmpty ParserError)

UnableToCompile MultipleErrors

renderBuildError :: BuildError -> String

buildSingleModule :: FilePath -> Text -> IO (Either BuildError (Module Ann))

Parse and build a single PureScript module, returning the compiled CoreFn module.

buildCoreFnOnly :: Env -> [ExternsFile] -> Module -> IO (Either MultipleErrors (Module Ann, MultipleErrors))

Compile a single Module into a CoreFn module.

This function is based on the rebuildModule function.

It is reproduced and modified here in order to make it simpler to build a single module without all of the additional capabilities and complexity of the upstream API.