Language.PureScript.Interpreter

High-level API

interpret :: FromValue m a => Env m -> Module ann -> EvalT m a

Evaluate a compiled PureScript module in the specified environment, returning the output from its main function as a Haskell value.

builtIn :: ToValue m a => Text -> a -> Env m

Create an environment from a Haskell value.

It is recommended that a type annotation is given for the type of the value being provided.

For example:

   builtIn "greeting" ("Hello, World!" :: Text)
   builtIn "somePrimes" ([2, 3, 5, 7, 11] :: Vector Integer)

Functions can be provided as built-ins, but the EvalT monad needs to be used to wrap any outputs (or values in positive position):

   builtIn "strip" ((pure . Text.strip) :: Text -> EvalT m Text)
   builtIn "map" (traverse :: (Value -> EvalT m Value) -> Vector Value -> EvalT m (Vector Value))

Polymorphic functions can also be provided as built-ins, but values with polymoprhic types will need to be passed across the FFI boundary with monomorphic types. The type Value can always be used to represent values of unknown or polymorphic type, as in the map example above.

Evaluation

Representation of values

data Value m

The representation of values used by the interpreter - essentially, the semantic domain for a simple untyped lambda calculus with records and ADTs.

Any additional side effects which might occur in FFI calls to Haskell code are tracked by a monad in the type argument.

Constructors

Object (HashMap Text (Value m))
Records are represented as hashmaps from their field names to values

Array (Vector (Value m))
Arrays

String Text

Number Scientific
Numeric values

TODO: separate integers from floating-point values

Bool Bool

Closure (Value m -> EvalT m (Value m))
Closures, represented in higher-order abstract syntax style.

Constructor (Qualified (ProperName 'ConstructorName)) [Value m]
Fully-applied data constructors

Evaluation monad

newtype EvalT m a

The monad used by the interpreter, which supports error reporting for errors which can occur during evaluation.

The transformed monad is used to track any benign side effects that might be exposed via the foreign function interface to PureScript code.

Constructors

EvalT

Fields

runEvalT :: ExceptT EvaluationError m a

type Eval = EvalT Identity

Non-transformer version of EvalT, useful in any settings where the FFI does not use any side effects during evaluation.

runEval :: Eval a -> Either EvaluationError a

data EvaluationError

Errors which can occur during evaluation of PureScript code.

PureScript is a typed language, and tries to prevent runtime errors. However, in the context of this interpreter, we can receive data from outside PureScript code, so it is possible that runtime errors can occur if we are not careful. This is similar to how PureScript code can fail at runtime due to errors in the FFI.

Constructors

UnknownIdent (Qualified Ident)
A name was not found in the environment

TypeMismatch Text
The runtime representation of a value did not match the expected representation

FieldNotFound Text
A record field did not exist in an Object value.

InexhaustivePatternMatch
A pattern match failed to match its argument

InvalidNumberOfArguments Int Int
A pattern match received the wrong number of arguments

UnsaturatedConstructorApplication
A pattern match occurred against a partially-applied data constructor

InvalidFieldName PSString
A PureScript string which contains lone surrogates which could not be decoded. See PSString.

NotSupported Text
A feature of the language is not yet implemented in the interpreter.

TODO: remove this once all features are supported

OtherError Text
An error occurred in a foreign function which is not tracked by any of the other error types.

TODO: remove this in favor of using monadic FFI functions

renderEvaluationError :: EvaluationError -> String

Render an EvaluationError as a human-readable string.

Eval/apply

type Env m = Map (Qualified Ident) (Value m)

An environment, i.e. a mapping from names to evaluated values.

An environment for a single built-in function can be constructed using the builtIn function, and environments can be combined easily using the Monoid instance for Map.

eval :: forall m. Monad m => ModuleName -> Env m -> Expr () -> EvalT m (Value m)

Evaluate a PureScript CoreFn expression in the given environment.

Note: it should not be necessary to call this function directly in most circumstances. It is provided as a helper function, for some more advanced use cases, such as setting up a custom environment.

apply :: Monad m => Value m -> Value m -> EvalT m (Value m)

Apply a value which represents an unevaluated closure to an argument.

Conversion to and from Haskell types

class Monad m => ToValue m a where

Minimal complete definition

Nothing

Methods

toValue :: a -> Value m

default toValue :: (Generic a, GToObject m (Rep a)) => a -> Value m

class Monad m => FromValue m a where

Minimal complete definition

Nothing

Methods

fromValue :: Value m -> EvalT m a

default fromValue :: (Generic a, GFromObject m (Rep a)) => Value m -> EvalT m a

Higher-order functions

class ToValueRHS m a where

Methods

toValueRHS :: a -> EvalT m (Value m)

class FromValueRHS m a where

Methods

fromValueRHS :: EvalT m (Value m) -> a

Records

class GToObject m f where

Methods

gToObject :: f x -> HashMap Text (Value m)

class GFromObject m f where

Methods

gFromObject :: HashMap Text (Value m) -> EvalT m (f x)