Language.PureScript.Interpreter

Documentation

module Language.PureScript.Interpreter.Types

module Language.PureScript.Interpreter.FFI

module Language.PureScript.Interpreter.FFI

module Language.PureScript.Interpreter.FFI.Builder

High-level API

data InterpretT m a

A monad transformer for high-level tasks involving PureScript code, including separate compilation. Its job is to keep track of available modules, any foreign imports from Haskell code, and run PureScript code.

Note: do not confuse this monad transformer with EvalT, which is only responsible for powering evaluation of PureScript expressions.

The transformed monad is used to track any benign side effects that might be exposed via the foreign function interface to PureScript code, in the same sense as EvalT.

runInterpretT :: Monad m => InterpretT m a -> m (Either InterpretError a)

Run a computation in the InterpretT monad, possibly returning an error. Note: errors can occur during module building or evaluation (i.e. module loading).

The runInterpret function is a simpler alternative in the case where benign side-effects are not needed.

For example:

runInterpret Module do
  -- Load the prelude
  ffi prelude
  -- Build a module from source
  build "module Main where main = \"example\"" --

runInterpret (Eval Text) do
  ffi prelude
  _ <- build "module Main where main = \"example\""
  -- Evaluate the main function
  evalMain (ModuleName "Main")

runInterpret :: Interpret a -> Either InterpretError a

Error messages

data InterpretError

The type of errors that can occur in the InterpretT monad.

Constructors

EvaluationError EvaluationError
Evaluation errors can occur during the initial evaluation of the module when it is loaded into the environment.

BuildError BuildError
Build errors can occur if we are building modules from source or corefn.

renderInterpretError :: InterpretError -> String

Foreign function interface

ffi :: Monad m => FFI m -> InterpretT m ()

Make an FFI module available for use to subsequent operations.

For example, to make the prelude available:

ffi prelude

Building PureScript source

build :: MonadFix m => Text -> InterpretT m (Module Ann)

Build a PureScript module from source, and make its exported functions available during subsequent evaluations.

buildCoreFn :: MonadFix m => Module Ann -> InterpretT m (Module Ann)

Build a PureScript module from corefn, and make its exported functions available during subsequent evaluations.

The corefn module may be preprepared, for example by compiling from source text using the functions in the Language.PureScript.Make.Simplified module.

Evaluating values

eval :: (MonadFix m, ToValueRHS m a) => Expr () -> InterpretT m a

Evaluate a PureScript corefn expression and return the result. Note: The expression is not type-checked by the PureScript typechecker. See the documentation for ToValueRHS for valid result types.

evalMain :: (MonadFix m, ToValueRHS m a) => ModuleName -> InterpretT m a

Evaluate main in the specified module and return the result.

type Eval = EvalT Identity

Non-transformer version of EvalT, useful in any settings where the FFI does not use any side effects during evaluation.

newtype EvalT m a

The monad used by the interpreter, which supports error reporting for errors which can occur during evaluation.

The transformed monad is used to track any benign side effects that might be exposed via the foreign function interface to PureScript code.

Constructors

EvalT

Fields

unEvalT :: ExceptT EvaluationError m a

type Env m = Map (Qualified Ident) (Value m)

An environment, i.e. a mapping from names to evaluated values.

An environment for a single built-in function can be constructed using the builtIn function, and environments can be combined easily using the Monoid instance for Map.

runEvalT :: EvalT m a -> m (Either EvaluationError a)

runEval :: Eval a -> Either EvaluationError a

class ToValueRHS m a where

ToValue should support functions with types such as

a -> EvalT m b
a -> b -> EvalT m c
a -> b -> c -> EvalT m d
(a -> EvalT m b) -> EvalT m c
(a -> b -> EvalT m c) -> EvalT m d

Note that every type in a return position is wrapped in the EvalT monad transformer. This is because evaluation in general may result in errors. However, a naive translation would result in too many applications of EvalT.

Specifically, we do not want to require types such as these, in which EvalT appears on the right hand side of every function arrow:

a -> EvalT m b (b -> EvalT m c)
a -> EvalT m b (b -> EvalT m (c -> EvalT m d))

For this reason, the ToValue instance for functions delegates to this type class for the type on the right hand side of the function. It skips the application of EvalT for nested function types.

Methods

toValueRHS :: a -> EvalT m (Value m)

fromValueRHS :: EvalT m (Value m) -> a

class MonadFix m => ToValue m a where

Values which can be communicated across the FFI boundary from Haskell to PureScript.

Instances should identify and document any valid representations as a subset of the semantic domain Value. Such a subset can be identified by an injective function toValue, and a partial inverse, fromValue, defined on the image of toValue.

Laws:

fromValue . toValue = pure

Minimal complete definition

Nothing

Methods

toValue :: a -> Value m

The default implementation uses generic deriving to identify a Haskell record type with a single data constructor with a PureScript record with the same field names.

default toValue :: (Generic a, ToObject m (Rep a)) => a -> Value m

fromValue :: Value m -> EvalT m a

default fromValue :: (Generic a, ToObject m (Rep a)) => Value m -> EvalT m a

data BuildError

Constructors

UnableToParse (NonEmpty ParserError)

UnableToCompile MultipleErrors

renderBuildError :: BuildError -> String

type Ann = (SourceSpan, [Comment], Maybe SourceType, Maybe Meta)

data Expr a

data Module a